# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kazaa API",
        "description": "API documentation for Kazaa system",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local server"
        }
    ],
    "paths": {
        "/api/en/user/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register new user and send OTP",
                "description": "Register user using name, phone number, country dial code and account type, then send OTP",
                "operationId": "registerUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "title": "CustomerRegister",
                                        "required": [
                                            "name",
                                            "country_dial_code_id",
                                            "phone_number",
                                            "account_type"
                                        ],
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Menna"
                                            },
                                            "country_dial_code_id": {
                                                "type": "integer",
                                                "example": 20
                                            },
                                            "phone_number": {
                                                "type": "string",
                                                "example": "01012345678"
                                            },
                                            "account_type": {
                                                "type": "string",
                                                "enum": [
                                                    "customer"
                                                ],
                                                "example": "customer"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    {
                                        "title": "VendorRepresentativeRegister",
                                        "required": [
                                            "name",
                                            "country_dial_code_id",
                                            "phone_number",
                                            "account_type",
                                            "email",
                                            "password",
                                            "position"
                                        ],
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Menna"
                                            },
                                            "country_dial_code_id": {
                                                "type": "integer",
                                                "example": 20
                                            },
                                            "phone_number": {
                                                "type": "string",
                                                "example": "01012345678"
                                            },
                                            "account_type": {
                                                "type": "string",
                                                "enum": [
                                                    "vendor_representative"
                                                ],
                                                "example": "vendor_representative"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "vendor@test.com"
                                            },
                                            "password": {
                                                "type": "string",
                                                "example": "123456"
                                            },
                                            "position": {
                                                "type": "string",
                                                "example": "Sales Manager"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Registered successfully, OTP sent to mobile"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/en/user/auth/verify-otp-register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Verifying OTP",
                "description": "Register user using  phone number, country dial code and OTP to verify his account",
                "operationId": "verify-otp-register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "country_dial_code_id",
                                    "phone_number",
                                    "otp"
                                ],
                                "properties": {
                                    "country_dial_code_id": {
                                        "description": "Country dial code ID",
                                        "type": "integer",
                                        "example": 242
                                    },
                                    "phone_number": {
                                        "description": "User mobile number",
                                        "type": "string",
                                        "example": "01012345678"
                                    },
                                    "otp": {
                                        "description": "OTP sent to mobile",
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "verification successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Registered successfully, OTP sent to mobile"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "success": true,
                                    "token": "3|QrRpUFZQO5KsBr0G1uONl3EftznaappuMcnuOlg79950949e",
                                    "data": {
                                        "user_id": 7,
                                        "user_name": "Menna",
                                        "user_slug": "menna",
                                        "user_email": "menna1@test.com",
                                        "user_country_dial_code_id": 242,
                                        "user_phone_number": "01096856825",
                                        "user_account_type": "vendor_representative",
                                        "vendor_representative": {
                                            "vendor_rep_id": 5,
                                            "vendor_rep_position": "manager"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "user not found"
                    },
                    "400": {
                        "description": "Invalid or expired OTP"
                    }
                }
            }
        },
        "http://127.0.0.1:8000/api/en/user/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout User",
                "description": "Logout User From All Devices",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Logged out successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Logged out successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "AppAPIKey": {
                "type": "apiKey",
                "name": "X-API-KEY",
                "in": "header"
            },
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}